fix(core): sync ProposalEffectVerifier tests with verifyAll() API
packages/core/__tests__/proposal-effect-verifier.test.ts was importing
EffectVerdict (removed) and calling verify() (renamed to verifyAll()),
causing TypeScript errors and test failures in CI.

Replace with the correct test suite that matches the current verifyAll()
API (adapted from src/__tests__/engine/ with updated import paths).
Stub out src/__tests__/engine/proposal-effect-verifier.test.ts per
codebase convention (real tests live in packages/core/__tests__/).
