+++ packages/core/__tests__/proposal-effect-verifier.test.ts
@@ -2,347 +2,437 @@
  * Tests for ProposalEffectVerifier (Spec 55 §7.7 Phase 2).
  *
  * Verifies verdict computation, signal writes, rollback_candidate emission,
- * outcome filtering, default signalRef parsing, and clock injection.
+ * outcome filtering, default signalRef parsing, and verifyThreshold.
  */
 
-import { beforeEach, describe, expect, it } from "bun:test";
+import { describe, expect, it } from "bun:test";
 import {
   createProposalEffectVerifier,
-  type EffectVerdict,
+  type EffectVerificationPayload,
   ProposalEffectVerifier,
 } from "../src/engine/proposal-effect-verifier";
 import type { ProposalOutcomePayload } from "../src/engine/proposal-outcome-recorder";
 import { InMemoryMemoryStore } from "../src/life-system/in-memory-memory-store";
-import type { Signal } from "../src/types/life-system";
+import type { Baseline, Signal } from "../src/types/life-system";
 
-// ── Fixtures ─────────────────────────────────────────────
+// ── Helpers ───────────────────────────────────────────────────────────────────
 
-function makeOutcome(overrides: Partial<ProposalOutcomePayload> = {}): ProposalOutcomePayload {
-  return {
-    proposalId: "proposal-abc",
+function makeStore(): InMemoryMemoryStore {
+  return new InMemoryMemoryStore();
+}
+
+function makeMergedSignal(overrides: Partial<ProposalOutcomePayload> = {}): Signal {
+  const payload: ProposalOutcomePayload = {
+    proposalId: "prop-xyz-00000001",
     capability: "cap-task",
     changeType: "minor",
     outcome: "merged",
-    recordedAt: "2026-01-01T00:00:00.000Z",
-    successMetric: {
-      signalRef: "task.completion_rate",
-      baselineValue: 0.5,
-      targetValue: 0.8,
-    },
+    recordedAt: new Date("2026-01-15T10:00:00Z").toISOString(),
     ...overrides,
+  };
+  return {
+    type: "proposal:outcome:merged",
+    source: "event_bus",
+    timestamp: new Date("2026-01-15T10:00:00Z"),
+    payload,
   };
 }
 
-async function lastSignal(store: InMemoryMemoryStore): Promise<Signal> {
-  const sigs = await store.getSignals();
-  const s = sigs.at(-1);
-  if (!s) throw new Error("No signal recorded");
-  return s;
+function makeBaseline(entity: string, metric: string, value: number): Baseline {
+  return { entity, metric, value, calculatedAt: new Date("2026-01-20T00:00:00Z") };
 }
 
-async function signalsByType(store: InMemoryMemoryStore, type: string): Promise<Signal[]> {
-  const sigs = await store.getSignals();
-  return sigs.filter((s) => s.type === type);
-}
-
-// ── Constructor and factory ───────────────────────────────
+// ── ProposalEffectVerifier — factory ─────────────────────────────────────────
 
 describe("ProposalEffectVerifier", () => {
-  let store: InMemoryMemoryStore;
-  let verifier: ProposalEffectVerifier;
-
-  beforeEach(() => {
-    store = new InMemoryMemoryStore();
-    verifier = new ProposalEffectVerifier({ store });
-  });
-
-  it("creates an instance via factory function", () => {
-    const v = createProposalEffectVerifier({ store });
-    expect(v).toBeInstanceOf(ProposalEffectVerifier);
-  });
-
-  // ── Filtering ─────────────────────────────────────────
-
-  it("skips non-merged outcomes by default", async () => {
-    const outcomes = [
-      makeOutcome({ outcome: "accepted" }),
-      makeOutcome({ outcome: "rejected" }),
-      makeOutcome({ outcome: "withdrawn" }),
-    ];
-    const results = await verifier.verify({ outcomes });
-    expect(results).toHaveLength(0);
-    expect(store.signalCount).toBe(0);
-  });
-
-  it("skips merged outcomes without successMetric", async () => {
-    const outcome = makeOutcome({ successMetric: undefined });
-    const results = await verifier.verify({ outcomes: [outcome] });
-    expect(results).toHaveLength(0);
-    expect(store.signalCount).toBe(0);
-  });
-
-  it("returns empty results for empty input", async () => {
-    const results = await verifier.verify({ outcomes: [] });
-    expect(results).toHaveLength(0);
-  });
-
-  it("applies custom outcomeFilter", async () => {
-    const outcomes = [makeOutcome({ outcome: "accepted" }), makeOutcome({ outcome: "merged" })];
-    // filter to "accepted" only
-    const results = await verifier.verify({ outcomes, outcomeFilter: "accepted" });
-    expect(results).toHaveLength(1);
-    expect(results[0].proposalId).toBe("proposal-abc");
-  });
-
-  // ── Verdict: effect_uncertain ─────────────────────────
-
-  it("returns effect_uncertain when signalRef is absent", async () => {
-    const outcome = makeOutcome({
-      successMetric: { signalRef: undefined, baselineValue: 0.5, targetValue: 0.8 },
+  describe("factory", () => {
+    it("createProposalEffectVerifier returns a ProposalEffectVerifier", () => {
+      const verifier = createProposalEffectVerifier({ store: makeStore() });
+      expect(verifier).toBeInstanceOf(ProposalEffectVerifier);
     });
-    const [result] = await verifier.verify({ outcomes: [outcome] });
-    expect(result.verdict).toBe("effect_uncertain");
-    expect(result.currentValue).toBeNull();
   });
 
-  it("returns effect_uncertain when getCurrentValue returns null", async () => {
-    const v = createProposalEffectVerifier({
-      store,
-      getCurrentValue: async () => null,
+  // ── verifyAll — skip conditions ───────────────────────────────────────────
+
+  describe("verifyAll — skip conditions", () => {
+    it("returns empty array when no merged outcome signals exist", async () => {
+      const store = makeStore();
+      const verifier = new ProposalEffectVerifier({ store });
+      expect(await verifier.verifyAll()).toEqual([]);
     });
-    const [result] = await v.verify({ outcomes: [makeOutcome()] });
-    expect(result.verdict).toBe("effect_uncertain");
-    expect(result.currentValue).toBeNull();
-  });
 
-  it("writes proposal:effect:uncertain signal on uncertainty", async () => {
-    const v = createProposalEffectVerifier({
-      store,
-      getCurrentValue: async () => null,
+    it("skips merged signals without successMetric", async () => {
+      const store = makeStore();
+      await store.recordSignal(makeMergedSignal()); // no successMetric
+      const verifier = new ProposalEffectVerifier({ store });
+      expect(await verifier.verifyAll()).toEqual([]);
     });
-    await v.verify({ outcomes: [makeOutcome()] });
-    const sigs = await signalsByType(store, "proposal:effect:uncertain");
-    expect(sigs).toHaveLength(1);
-    const payload = sigs[0].payload as { verdict: EffectVerdict };
-    expect(payload.verdict).toBe("effect_uncertain");
-  });
 
-  // ── Verdict: effect_verified ──────────────────────────
-
-  it("returns effect_verified when currentValue >= targetValue", async () => {
-    const v = createProposalEffectVerifier({
-      store,
-      getCurrentValue: async () => 0.9,
-    });
-    const [result] = await v.verify({ outcomes: [makeOutcome()] });
-    expect(result.verdict).toBe("effect_verified");
-    expect(result.currentValue).toBe(0.9);
-  });
-
-  it("returns effect_verified when currentValue exactly equals targetValue", async () => {
-    const v = createProposalEffectVerifier({
-      store,
-      getCurrentValue: async () => 0.8, // exactly equal to targetValue: 0.8
-    });
-    const [result] = await v.verify({ outcomes: [makeOutcome()] });
-    expect(result.verdict).toBe("effect_verified");
-  });
-
-  it("writes proposal:effect:verified signal on success", async () => {
-    const v = createProposalEffectVerifier({
-      store,
-      getCurrentValue: async () => 0.9,
-    });
-    await v.verify({ outcomes: [makeOutcome()] });
-    const sigs = await signalsByType(store, "proposal:effect:verified");
-    expect(sigs).toHaveLength(1);
-  });
-
-  it("does NOT write rollback_candidate on effect_verified", async () => {
-    const v = createProposalEffectVerifier({
-      store,
-      getCurrentValue: async () => 0.9,
-    });
-    await v.verify({ outcomes: [makeOutcome()] });
-    const rbSigs = await signalsByType(store, "proposal:effect:rollback_candidate");
-    expect(rbSigs).toHaveLength(0);
-  });
-
-  // ── Verdict: effect_failed ────────────────────────────
-
-  it("returns effect_failed when currentValue < targetValue", async () => {
-    const v = createProposalEffectVerifier({
-      store,
-      getCurrentValue: async () => 0.4,
-    });
-    const [result] = await v.verify({ outcomes: [makeOutcome()] });
-    expect(result.verdict).toBe("effect_failed");
-    expect(result.currentValue).toBe(0.4);
-  });
-
-  it("writes proposal:effect:failed signal on failure", async () => {
-    const v = createProposalEffectVerifier({
-      store,
-      getCurrentValue: async () => 0.4,
-    });
-    await v.verify({ outcomes: [makeOutcome()] });
-    const sigs = await signalsByType(store, "proposal:effect:failed");
-    expect(sigs).toHaveLength(1);
-  });
-
-  it("writes rollback_candidate signal on effect_failed", async () => {
-    const v = createProposalEffectVerifier({
-      store,
-      getCurrentValue: async () => 0.4,
-    });
-    await v.verify({ outcomes: [makeOutcome()] });
-    const rbSigs = await signalsByType(store, "proposal:effect:rollback_candidate");
-    expect(rbSigs).toHaveLength(1);
-    const payload = rbSigs[0].payload as { verdict: EffectVerdict; proposalId: string };
-    expect(payload.verdict).toBe("effect_failed");
-    expect(payload.proposalId).toBe("proposal-abc");
-  });
-
-  it("writes 2 signals total on effect_failed (failed + rollback_candidate)", async () => {
-    const v = createProposalEffectVerifier({
-      store,
-      getCurrentValue: async () => 0.4,
-    });
-    await v.verify({ outcomes: [makeOutcome()] });
-    expect(store.signalCount).toBe(2);
-  });
-
-  // ── Multiple outcomes ─────────────────────────────────
-
-  it("processes multiple eligible outcomes independently", async () => {
-    const outcomes = [
-      makeOutcome({
-        proposalId: "p1",
-        successMetric: { baselineValue: 0, targetValue: 0.8, signalRef: "a.b" },
-      }),
-      makeOutcome({
-        proposalId: "p2",
-        successMetric: { baselineValue: 0, targetValue: 0.5, signalRef: "c.d" },
-      }),
-    ];
-    const calls: string[] = [];
-    const v = createProposalEffectVerifier({
-      store,
-      getCurrentValue: async (ref) => {
-        calls.push(ref);
-        return ref === "a.b" ? 0.9 : 0.2;
-      },
-    });
-    const results = await v.verify({ outcomes });
-    expect(results).toHaveLength(2);
-    expect(results[0].verdict).toBe("effect_verified");
-    expect(results[1].verdict).toBe("effect_failed");
-    expect(calls).toEqual(["a.b", "c.d"]);
-  });
-
-  // ── Result payload ────────────────────────────────────
-
-  it("result includes proposalId, capability, changeType, successMetric snapshot", async () => {
-    const v = createProposalEffectVerifier({
-      store,
-      getCurrentValue: async () => 0.9,
-    });
-    const [result] = await v.verify({ outcomes: [makeOutcome()] });
-    expect(result.proposalId).toBe("proposal-abc");
-    expect(result.capability).toBe("cap-task");
-    expect(result.changeType).toBe("minor");
-    expect(result.successMetric.baselineValue).toBe(0.5);
-    expect(result.successMetric.targetValue).toBe(0.8);
-    expect(result.successMetric.signalRef).toBe("task.completion_rate");
-  });
-
-  it("result includes verifiedAt ISO 8601 timestamp", async () => {
-    const fixedNow = new Date("2026-06-01T12:00:00.000Z");
-    const v = createProposalEffectVerifier({
-      store,
-      getCurrentValue: async () => 0.9,
-      clock: () => fixedNow,
-    });
-    const [result] = await v.verify({ outcomes: [makeOutcome()] });
-    expect(result.verifiedAt).toBe("2026-06-01T12:00:00.000Z");
-  });
-
-  it("signal timestamp matches clock injection", async () => {
-    const fixedNow = new Date("2026-06-01T12:00:00.000Z");
-    const v = createProposalEffectVerifier({
-      store,
-      getCurrentValue: async () => 0.9,
-      clock: () => fixedNow,
-    });
-    await v.verify({ outcomes: [makeOutcome()] });
-    const sig = await lastSignal(store);
-    expect(sig.timestamp).toEqual(fixedNow);
-  });
-
-  // ── Default getCurrentValue (baseline lookup) ─────────
-
-  it("default getCurrentValue parses dot-separated signalRef (entity.metric)", async () => {
-    await store.updateBaseline({
-      entity: "task",
-      metric: "completion_rate",
-      value: 0.9,
-      calculatedAt: new Date(),
-    });
-    const [result] = await verifier.verify({
-      outcomes: [
-        makeOutcome({
-          successMetric: {
-            signalRef: "task.completion_rate",
-            baselineValue: 0.5,
-            targetValue: 0.8,
-          },
+    it("skips merged signals with successMetric but without signalRef", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          successMetric: { baselineValue: 10, targetValue: 20 }, // no signalRef
         }),
-      ],
+      );
+      const verifier = new ProposalEffectVerifier({ store });
+      expect(await verifier.verifyAll()).toEqual([]);
     });
-    expect(result.currentValue).toBe(0.9);
-    expect(result.verdict).toBe("effect_verified");
+
+    it("does not process non-merged outcome signals", async () => {
+      const store = makeStore();
+      await store.recordSignal({
+        type: "proposal:outcome:accepted",
+        source: "event_bus",
+        timestamp: new Date(),
+        payload: {
+          proposalId: "prop-a",
+          capability: "cap-x",
+          changeType: "minor",
+          outcome: "accepted",
+          recordedAt: new Date().toISOString(),
+          successMetric: { baselineValue: 5, targetValue: 50, signalRef: "cap-x" },
+        } satisfies ProposalOutcomePayload,
+      });
+      const verifier = new ProposalEffectVerifier({ store });
+      expect(await verifier.verifyAll()).toEqual([]);
+    });
   });
 
-  it("default getCurrentValue parses colon-separated signalRef (entity:metric)", async () => {
-    await store.updateBaseline({
-      entity: "order",
-      metric: "error_rate",
-      value: 0.1,
-      calculatedAt: new Date(),
-    });
-    const [result] = await verifier.verify({
-      outcomes: [
-        makeOutcome({
-          successMetric: { signalRef: "order:error_rate", baselineValue: 0.3, targetValue: 0.05 },
+  // ── verifyAll — effect_uncertain (no baseline) ────────────────────────────
+
+  describe("verifyAll — effect_uncertain (no baseline)", () => {
+    it("returns effect_uncertain when signalRef has no stored baseline", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          successMetric: { baselineValue: 10, targetValue: 80, signalRef: "cap-task" },
         }),
-      ],
+      );
+      const verifier = new ProposalEffectVerifier({ store });
+      const [rec] = await verifier.verifyAll();
+
+      expect(rec.result).toBe("effect_uncertain");
+      expect(rec.currentValue).toBeUndefined();
+      expect(rec.proposalId).toBe("prop-xyz-00000001");
     });
-    // currentValue=0.1 >= targetValue=0.05 → effect_verified
-    expect(result.currentValue).toBe(0.1);
-    expect(result.verdict).toBe("effect_verified");
+
+    it("emits proposal:effect:uncertain signal to store", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          successMetric: { baselineValue: 10, targetValue: 80, signalRef: "cap-task" },
+        }),
+      );
+      const verifier = new ProposalEffectVerifier({ store });
+      await verifier.verifyAll();
+
+      const effectSignals = await store.getSignals({ entity: "proposal:effect:uncertain" });
+      expect(effectSignals).toHaveLength(1);
+      const p = effectSignals[0].payload as EffectVerificationPayload;
+      expect(p.rollback_candidate).toBe(false);
+    });
   });
 
-  it("default getCurrentValue returns null when signalRef has no separator", async () => {
-    const [result] = await verifier.verify({
-      outcomes: [
-        makeOutcome({
-          successMetric: { signalRef: "noseparator", baselineValue: 0, targetValue: 1 },
+  // ── verifyAll — effect_uncertain (partial progress) ───────────────────────
+
+  describe("verifyAll — effect_uncertain (partial progress)", () => {
+    it("returns effect_uncertain when current is between baseline and 90% target", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          successMetric: { baselineValue: 0, targetValue: 100, signalRef: "cap-task:success_rate" },
         }),
-      ],
+      );
+      await store.updateBaseline(makeBaseline("cap-task", "success_rate", 50));
+      const verifier = new ProposalEffectVerifier({ store });
+      const [rec] = await verifier.verifyAll();
+
+      expect(rec.result).toBe("effect_uncertain");
+      expect(rec.currentValue).toBe(50);
     });
-    expect(result.currentValue).toBeNull();
-    expect(result.verdict).toBe("effect_uncertain");
   });
 
-  it("default getCurrentValue returns null when baseline not in store", async () => {
-    const [result] = await verifier.verify({
-      outcomes: [
-        makeOutcome({
-          successMetric: { signalRef: "entity.missing_metric", baselineValue: 0, targetValue: 1 },
+  // ── verifyAll — effect_verified ───────────────────────────────────────────
+
+  describe("verifyAll — effect_verified", () => {
+    it("returns effect_verified when current >= 90% of target gap (default threshold)", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          successMetric: { baselineValue: 0, targetValue: 100, signalRef: "cap-task" },
         }),
-      ],
+      );
+      await store.updateBaseline(makeBaseline("cap-task", "value", 95));
+      const verifier = new ProposalEffectVerifier({ store });
+      const [rec] = await verifier.verifyAll();
+
+      expect(rec.result).toBe("effect_verified");
+      expect(rec.currentValue).toBe(95);
     });
-    expect(result.currentValue).toBeNull();
-    expect(result.verdict).toBe("effect_uncertain");
+
+    it("returns effect_verified exactly at the threshold boundary", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          successMetric: { baselineValue: 0, targetValue: 100, signalRef: "cap-task" },
+        }),
+      );
+      await store.updateBaseline(makeBaseline("cap-task", "value", 90));
+      const verifier = new ProposalEffectVerifier({ store });
+      const [rec] = await verifier.verifyAll();
+
+      expect(rec.result).toBe("effect_verified");
+    });
+
+    it("emits proposal:effect:verified signal with rollback_candidate: false", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          successMetric: { baselineValue: 0, targetValue: 100, signalRef: "cap-task" },
+        }),
+      );
+      await store.updateBaseline(makeBaseline("cap-task", "value", 95));
+      const verifier = new ProposalEffectVerifier({ store });
+      await verifier.verifyAll();
+
+      const effectSignals = await store.getSignals({ entity: "proposal:effect:verified" });
+      expect(effectSignals).toHaveLength(1);
+      const p = effectSignals[0].payload as EffectVerificationPayload;
+      expect(p.rollback_candidate).toBe(false);
+    });
+
+    it("respects custom verifyThreshold", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          successMetric: { baselineValue: 0, targetValue: 100, signalRef: "cap-task" },
+        }),
+      );
+      await store.updateBaseline(makeBaseline("cap-task", "value", 70));
+      const verifier = new ProposalEffectVerifier({ store, verifyThreshold: 0.6 });
+      const [rec] = await verifier.verifyAll();
+
+      expect(rec.result).toBe("effect_verified");
+    });
+  });
+
+  // ── verifyAll — effect_failed ─────────────────────────────────────────────
+
+  describe("verifyAll — effect_failed", () => {
+    it("returns effect_failed when current <= baselineValue", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          successMetric: { baselineValue: 50, targetValue: 80, signalRef: "cap-task" },
+        }),
+      );
+      await store.updateBaseline(makeBaseline("cap-task", "value", 45));
+      const verifier = new ProposalEffectVerifier({ store });
+      const [rec] = await verifier.verifyAll();
+
+      expect(rec.result).toBe("effect_failed");
+      expect(rec.currentValue).toBe(45);
+    });
+
+    it("returns effect_failed when current equals baselineValue", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          successMetric: { baselineValue: 50, targetValue: 80, signalRef: "cap-task" },
+        }),
+      );
+      await store.updateBaseline(makeBaseline("cap-task", "value", 50));
+      const verifier = new ProposalEffectVerifier({ store });
+      const [rec] = await verifier.verifyAll();
+
+      expect(rec.result).toBe("effect_failed");
+    });
+
+    it("emits proposal:effect:failed signal with rollback_candidate: true", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          successMetric: { baselineValue: 50, targetValue: 80, signalRef: "cap-task" },
+        }),
+      );
+      await store.updateBaseline(makeBaseline("cap-task", "value", 40));
+      const verifier = new ProposalEffectVerifier({ store });
+      await verifier.verifyAll();
+
+      const failedSignals = await store.getSignals({ entity: "proposal:effect:failed" });
+      expect(failedSignals).toHaveLength(1);
+      const p = failedSignals[0].payload as EffectVerificationPayload;
+      expect(p.rollback_candidate).toBe(true);
+      expect(p.proposalId).toBe("prop-xyz-00000001");
+    });
+  });
+
+  // ── signalRef parsing ─────────────────────────────────────────────────────
+
+  describe("signalRef parsing", () => {
+    it("parses bare entity ref (defaults metric to 'value')", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          successMetric: { baselineValue: 0, targetValue: 100, signalRef: "my-cap" },
+        }),
+      );
+      await store.updateBaseline(makeBaseline("my-cap", "value", 95));
+      const verifier = new ProposalEffectVerifier({ store });
+      const [rec] = await verifier.verifyAll();
+
+      expect(rec.result).toBe("effect_verified");
+      expect(rec.signalRef).toBe("my-cap");
+    });
+
+    it("parses entity:metric ref correctly", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          successMetric: { baselineValue: 0, targetValue: 100, signalRef: "cap-orders:fill_rate" },
+        }),
+      );
+      await store.updateBaseline(makeBaseline("cap-orders", "fill_rate", 95));
+      const verifier = new ProposalEffectVerifier({ store });
+      const [rec] = await verifier.verifyAll();
+
+      expect(rec.result).toBe("effect_verified");
+      expect(rec.signalRef).toBe("cap-orders:fill_rate");
+    });
+
+    it("treats only the first colon as the entity/metric separator", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          successMetric: { baselineValue: 0, targetValue: 100, signalRef: "cap-foo:bar:baz" },
+        }),
+      );
+      await store.updateBaseline(makeBaseline("cap-foo", "bar:baz", 95));
+      const verifier = new ProposalEffectVerifier({ store });
+      const [rec] = await verifier.verifyAll();
+
+      expect(rec.result).toBe("effect_verified");
+    });
+  });
+
+  // ── verifyAll — multiple proposals ───────────────────────────────────────
+
+  describe("verifyAll — multiple proposals", () => {
+    it("processes multiple merged signals and returns one record each", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          proposalId: "prop-a",
+          successMetric: { baselineValue: 0, targetValue: 100, signalRef: "cap-a" },
+        }),
+      );
+      await store.recordSignal(
+        makeMergedSignal({
+          proposalId: "prop-b",
+          successMetric: { baselineValue: 0, targetValue: 100, signalRef: "cap-b" },
+        }),
+      );
+      await store.updateBaseline(makeBaseline("cap-a", "value", 95));
+      const verifier = new ProposalEffectVerifier({ store });
+      const results = await verifier.verifyAll();
+
+      expect(results).toHaveLength(2);
+      const propA = results.find((r) => r.proposalId === "prop-a");
+      const propB = results.find((r) => r.proposalId === "prop-b");
+      expect(propA?.result).toBe("effect_verified");
+      expect(propB?.result).toBe("effect_uncertain");
+    });
+
+    it("emits one verification signal per processed proposal", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          proposalId: "prop-1",
+          successMetric: { baselineValue: 0, targetValue: 100, signalRef: "cap-x" },
+        }),
+      );
+      await store.recordSignal(
+        makeMergedSignal({
+          proposalId: "prop-2",
+          successMetric: { baselineValue: 0, targetValue: 100, signalRef: "cap-y" },
+        }),
+      );
+      await store.updateBaseline(makeBaseline("cap-x", "value", 95));
+      await store.updateBaseline(makeBaseline("cap-y", "value", 0));
+
+      const verifier = new ProposalEffectVerifier({ store });
+      await verifier.verifyAll();
+
+      const verifiedSigs = await store.getSignals({ entity: "proposal:effect:verified" });
+      const failedSigs = await store.getSignals({ entity: "proposal:effect:failed" });
+      expect(verifiedSigs).toHaveLength(1);
+      expect(failedSigs).toHaveLength(1);
+    });
+  });
+
+  // ── verifyAll — since filter ──────────────────────────────────────────────
+
+  describe("verifyAll — since filter", () => {
+    it("only processes signals recorded at or after the since date", async () => {
+      const store = makeStore();
+      await store.recordSignal({
+        type: "proposal:outcome:merged",
+        source: "event_bus",
+        timestamp: new Date("2026-01-01T00:00:00Z"),
+        payload: {
+          proposalId: "prop-old",
+          capability: "cap-z",
+          changeType: "minor",
+          outcome: "merged",
+          recordedAt: new Date("2026-01-01T00:00:00Z").toISOString(),
+          successMetric: { baselineValue: 0, targetValue: 100, signalRef: "cap-z" },
+        } satisfies ProposalOutcomePayload,
+      });
+      await store.recordSignal({
+        type: "proposal:outcome:merged",
+        source: "event_bus",
+        timestamp: new Date("2026-02-01T00:00:00Z"),
+        payload: {
+          proposalId: "prop-new",
+          capability: "cap-z",
+          changeType: "minor",
+          outcome: "merged",
+          recordedAt: new Date("2026-02-01T00:00:00Z").toISOString(),
+          successMetric: { baselineValue: 0, targetValue: 100, signalRef: "cap-z" },
+        } satisfies ProposalOutcomePayload,
+      });
+
+      const verifier = new ProposalEffectVerifier({ store });
+      const results = await verifier.verifyAll({ since: new Date("2026-01-20T00:00:00Z") });
+
+      expect(results).toHaveLength(1);
+      expect(results[0].proposalId).toBe("prop-new");
+    });
+  });
+
+  // ── EffectVerificationRecord fields ──────────────────────────────────────
+
+  describe("EffectVerificationRecord fields", () => {
+    it("record contains proposalId, capability, signalRef, baselineValue, targetValue, verifiedAt", async () => {
+      const store = makeStore();
+      await store.recordSignal(
+        makeMergedSignal({
+          proposalId: "prop-abc",
+          capability: "cap-task",
+          successMetric: { baselineValue: 10, targetValue: 90, signalRef: "cap-task:rate" },
+        }),
+      );
+      await store.updateBaseline(makeBaseline("cap-task", "rate", 85));
+      const verifier = new ProposalEffectVerifier({ store });
+      const [rec] = await verifier.verifyAll();
+
+      expect(rec.proposalId).toBe("prop-abc");
+      expect(rec.capability).toBe("cap-task");
+      expect(rec.signalRef).toBe("cap-task:rate");
+      expect(rec.baselineValue).toBe(10);
+      expect(rec.targetValue).toBe(90);
+      expect(rec.verifiedAt).toMatch(/^\d{4}-\d{2}-\d{2}T/);
+    });
   });
 });
